Add CI configuration for building and publishing docs
This commit adds the basic github action workflows to setup a docs
build and publish job. On every PR we will do a clean docs build
and upload the docs to the artifacts store for local inspection. Then on
every merge to master the docs will be built and uploaded to IBM COS for
hosting at https://qiskit.org/documentation/partners